pyproject.toml
[project]
name = "python-form"
version = "0.1.0"
description = "Dynamic YAML Form Generator"
readme = "README.md"
authors = [
    { name = "Nikolay Borissov", email = "nikolay.borissov@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.3.2",
    "pyyaml>=6.0.2",
    "streamlit>=1.48.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
